requirements.txt
streamlit
mediapipe==0.10.9
opencv-python
numpy
pandas
scikit-learn
joblib
Pillow